requirements.txt
future==1.0.0
iso8601==2.1.0
MouseInfo==0.1.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
# pyobjc-core==10.3.2
# pyobjc-framework-Cocoa==10.3.2
# pyobjc-framework-Quartz==10.3.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pytweening==1.2.0
PyYAML==6.0.2
rubicon-objc==0.5.0
